+++ src/app/(sidebar)/godplaces/page.tsx
@@ -1,4 +1,4 @@
-import Search from '@/components/godplaces/Search';
+import Search from '@/components/godplaces/search/Search';
 
 export default function GodPlaces() {
   return (
