+++ src/components/godplaces/detail/DetailBox.tsx
