+++ src/components/godplaces/detail/DetailCard.tsx
