+++ src/components/godplaces/detail/DetailTopInfo.tsx
