+++ src/components/godplaces/detail/DetailsItem.tsx
