+++ src/components/godplaces/detail/FestivalDetails.tsx
