+++ src/components/godplaces/detail/LibraryDetails.tsx
